+++ fixbackend/cloud_accounts/service_impl.py
@@ -57,7 +57,7 @@ from fixbackend.ids import (
     WorkspaceId,
 )
 from fixbackend.logging_context import set_cloud_account_id, set_fix_cloud_account_id, set_workspace_id
-from fixbackend.notification.messages import SecurityScanFinished
+from fixbackend.notification.email.email_messages import SecurityScanFinished
 from fixbackend.notification.service import NotificationService
 from fixbackend.sqs import SQSRawListener
 from fixbackend.utils import uid
