+++ fixbackend/notification/email/templates/invite.html
