+++ fixbackend/workspaces/invitation_service.py
@@ -31,7 +31,7 @@ from fixbackend.domain_events.events import InvitationAccepted
 from fixbackend.domain_events.publisher import DomainEventPublisher
 from fixbackend.ids import InvitationId, WorkspaceId
 from fixbackend.notification.service import NotificationService, EmailServiceDependency
-from fixbackend.notification.messages import Invite
+from fixbackend.notification.email.email_messages import Invite
 from fixbackend.workspaces.invitation_repository import InvitationRepository, InvitationRepositoryDependency
 from fixbackend.workspaces.models import WorkspaceInvitation
 from fixbackend.workspaces.repository import WorkspaceRepository, WorkspaceRepositoryDependency
